requirements.txt
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
fonttools==4.56.0
fredapi==0.5.2
frozendict==2.4.6
idna==3.10
joblib==1.4.2
kiwisolver==1.4.8
matplotlib==3.10.0
multitasking==0.0.11
narwhals==1.27.1
numpy==2.2.3
packaging==24.2
pandas==2.2.3
peewee==3.17.9
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
six==1.17.0
soupsieve==2.6
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
yfinance==0.2.54
